Write bigger chunks to ram_file
`ram_file` is implemented as a port driver, and has quiet a
significant overhead if every single series is emitted separately.

What happens here is that bigger chunk is collected before writing to
ram_file. Chunk is collected using binary append, in such a way that
is can benefit from binary append optimization in ERTS.

One nice side-effect of collecting binaries is that a lot of iolists
in the hottest code path are no longer created, putting significantly
less pressure on the garbage collector.