Allow application to pre-render labels for text format
Builds on top of https://github.com/deadtrickster/prometheus.erl/pull/136

This is useful when e.g. escaping is not needed, as it is in case with
RabbitMQ, where no escaping is needed because strict rules already
apply to names.

This change provides not only direct speedup, but also significantly
reduced garbage collection overhead, as for every output line the
proplist with labels and the list with converted `#'LabelPair'{}`'s
are replaced with just a single binary.

Using RabbitMQ as an example, time to produce ~630k metrics went from
~6 to ~3 seconds.

Downside is that it's only compatible with text format, and spec for
`#'LabelPair'{}' is not respected.