Support global and prerendered labels together (#149)
Without this change, a collector that pre-renders labels (like RabbitMQ)
will return `<<>>` in case there are no labels at all. That will create an
an improper list when an empty binary is appended to a list of global
labels. Rather than dealing with improper list later, let's just avoid
them.

Co-authored-by: Michal Kuratczyk <mkuratczyk@vmware.com>